+++ R/standardize_info.R
@@ -10,7 +10,7 @@
 #' model <- lm(Sepal.Width ~ Sepal.Length * Species, data = iris)
 #' @importFrom parameters parameters_type
 #' @export
-standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...) {
+standardize_info <- function(model, robust = FALSE, two_sd = FALSE, include_pseudo = FALSE, ...) {
   params <- insight::find_parameters(model, effects = "fixed", flatten = TRUE, ...)
   types <- parameters::parameters_type(model)
   model_matrix <- as.data.frame(stats::model.matrix(model))
@@ -57,14 +57,14 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
   # Basic
   out <- merge(
     out,
-    .std_info_predictors_basic(model_matrix, types, robust = robust),
+    .std_info_predictors_basic(model_matrix, types, robust = robust, two_sd = two_sd),
     by = "Parameter", all = TRUE
   )
 
   # Smart
   out <- merge(
     out,
-    .std_info_predictors_smart(data, params, types, robust = robust),
+    .std_info_predictors_smart(data, params, types, robust = robust, two_sd = two_sd),
     by = "Parameter", all = TRUE
   )
 
@@ -74,7 +74,7 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
       length(insight::find_random(model)$random) == 1) {
     out <- merge(
       out,
-      .std_info_pseudo(model, params, model_matrix, types = types$Type, robust = robust)
+      .std_info_pseudo(model, params, model_matrix, types = types$Type, robust = robust, two_sd = two_sd)
     )
   }
 
@@ -105,7 +105,7 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
 
 
 #' @keywords internal
-.std_info_predictors_smart <- function(data, params, types, robust = FALSE, ...) {
+.std_info_predictors_smart <- function(data, params, types, robust = FALSE, two_sd = FALSE, ...) {
 
   # Get deviations for all parameters
   means <- deviations <- rep(NA_real_, times = length(params))
@@ -115,7 +115,8 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
       data = data,
       variable = types[types$Parameter == var, "Variable"],
       type = types[types$Parameter == var, "Type"],
-      robust = robust
+      robust = robust,
+      two_sd = two_sd
     )
     deviations[i] <- info$sd
     means[i] <- info$mean
@@ -132,11 +133,11 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
 
 
 #' @keywords internal
-.std_info_predictor_smart <- function(data, variable, type, robust = FALSE, ...) {
+.std_info_predictor_smart <- function(data, variable, type, robust = FALSE, two_sd = FALSE, ...) {
   if (type == "intercept") {
     info <- list(sd = 0, mean = 0)
   } else if (type == "numeric") {
-    info <- .compute_std_info(data = data, variable = variable, robust = robust)
+    info <- .compute_std_info(data = data, variable = variable, robust = robust, two_sd = two_sd)
   } else if (type == "factor") {
     info <- list(sd = 1, mean = 0)
 
@@ -152,7 +153,7 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
     # }
   } else if (type %in% c("interaction", "nested")) {
     if (is.numeric(data[, variable])) {
-      info <- .compute_std_info(data = data, variable = variable, robust = robust)
+      info <- .compute_std_info(data = data, variable = variable, robust = robust, two_sd = two_sd)
     } else if (is.factor(data[, variable])) {
       info <- list(sd = 1, mean = 0)
     } else {
@@ -170,7 +171,7 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
 
 
 #' @keywords internal
-.std_info_predictors_basic <- function(model_matrix, types, robust = FALSE, ...) {
+.std_info_predictors_basic <- function(model_matrix, types, robust = FALSE, two_sd = FALSE, ...) {
 
   # Get deviations for all parameters
   means <- deviations <- rep(NA_real_, length = length(names(model_matrix)))
@@ -179,7 +180,7 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
     if (types[i, "Type"] == "intercept") {
       means[i] <- deviations[i] <- 0
     } else {
-      std_info <- .compute_std_info(data = model_matrix, variable = var, robust = robust)
+      std_info <- .compute_std_info(data = model_matrix, variable = var, robust = robust, two_sd = two_sd)
       deviations[i] <- std_info$sd
       means[i] <- std_info$mean
     }
@@ -268,11 +269,13 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
 #' @importFrom insight clean_names get_random model_info find_formula get_variance get_data
 #' @importFrom parameters check_heterogeneity demean
 #' @importFrom stats as.formula sd
-.std_info_pseudo <- function(model, params, model_matrix, types, robust = FALSE) {
+.std_info_pseudo <- function(model, params, model_matrix, types, robust = FALSE, two_sd = FALSE) {
   if (robust) {
     warning("'robust' standardization not available for 'pseudo' method.",
             call. = FALSE)
   }
+
+  f <- if (two_sd) 2 else 1
 
   within_vars <- unclass(parameters::check_heterogeneity(model))
   id <- insight::get_random(model)[[1]]
@@ -338,11 +341,11 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
     rand_name <- insight::find_random(model)$random
 
     # maintain any y-transformations
-    f <- insight::find_formula(model)
-    f <- paste0(f$conditional[2], " ~ (1|",rand_name,")")
+    frm <- insight::find_formula(model)
+    frm <- paste0(frm$conditional[2], " ~ (1|",rand_name,")")
 
     m0 <- suppressWarnings(suppressMessages(
-      lme4::lmer(stats::as.formula(f),
+      lme4::lmer(stats::as.formula(frm),
                  data = insight::get_data(model))
     ))
     m0v <- insight::get_variance(m0)
@@ -368,10 +371,10 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
         X <- tapply(model_matrix[[i]], id, mean)
         Deviation_Response_Pseudo[i] <- sd_y_between
       }
-      Deviation_Pseudo[i] <- stats::sd(X)
+      Deviation_Pseudo[i] <- f * stats::sd(X)
 
       ## smart way?
-      ## DONT USE: see corespondance with between Mattan and Eran BC
+      ## DONT USE: see correspondence with between Mattan and Eran BC
       # m <- suppressWarnings(suppressMessages(lme4::lmer(model_matrix[[i]] ~ (1|id))))
       # if (is_within[i]) {
       #   ## is within
@@ -402,7 +405,8 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
 
 
 #' @keywords internal
-.compute_std_info <- function(data = NULL, variable = NULL, response = NULL, robust = FALSE) {
+.compute_std_info <- function(data = NULL, variable = NULL, response = NULL, robust = FALSE, two_sd = FALSE) {
+  f <- if (two_sd) 2 else 1
   if (is.null(response)) {
     response <- as.numeric(data[, variable])
   }
@@ -415,5 +419,5 @@ standardize_info <- function(model, robust = FALSE, include_pseudo = FALSE, ...)
     mean_x <- stats::median(response, na.rm = TRUE)
   }
 
-  list(sd = sd_x, mean = mean_x)
+  list(sd = f * sd_x, mean = mean_x)
 }